fix(census): env kill switch covers relay stamps; test file split
Codex P2 x2: (1) HA_NOVA_NO_CENSUS now also suppresses the passive
relay-version accrual for opted-in installs (test added); (2)
census_test.go split at the carrier/inertness boundary into
census_carrier_test.go — both files back under the ~400 LOC
guideline.
